Refactor tests slightly and add coverage for all the connect methods, binary frames, and ping intervals.
